requirements.txt
fastapi
uvicorn[standard]
python-pptx
aiofiles
python-multipart